Move setting types to own file
